requirements.txt
streamlit
pillow
opencv-python-headless
torch
transformers
moviepy
numpy
